add CSV package to data.source and move project, task, and user packages into it
