+++ src/main/kotlin/data/repository/ProjectRepositoryImpl.kt
@@ -2,8 +2,8 @@ package data.repository
 
 import data.mapper.toDomain
 import data.mapper.toDto
-import data.source.project.ProjectExternalDataSource
-import data.source.user.CurrentUserProvider
+import data.source.csv.project.ProjectExternalDataSource
+import data.source.csv.user.CurrentUserProvider
 import data.utils.toProjectException
 import domain.models.project.Project
 import domain.repository.ProjectRepository
