+++ src/main/kotlin/data/source/csv/user/UserMemoryDataSource.kt
@@ -1,4 +1,4 @@
-package data.source.user
+package data.source.csv.user
 
 import data.dto.authentication.UserDto
