+++ src/main/kotlin/di/modules/repositoryModule.kt
@@ -3,7 +3,7 @@ package di.modules
 import data.repository.ProjectRepositoryImpl
 import data.repository.TaskRepositoryImpl
 import data.repository.UserRepositoryImpl
-import data.source.project.ProjectManager
+import data.source.csv.project.ProjectManager
 import domain.repository.ProjectRepository
 import domain.repository.TaskRepository
 import domain.repository.UserRepository
