+++ src/test/kotlin/data/repository/ProjectRepositoryImplTest.kt
@@ -4,9 +4,9 @@ import com.google.common.truth.Truth.assertThat
 import data.createProject
 import data.createUserDto
 import data.mapper.toDto
-import data.source.project.ProjectExternalDataSource
-import data.source.project.ProjectManager
-import data.source.user.CurrentUserProvider
+import data.source.csv.project.ProjectExternalDataSource
+import data.source.csv.project.ProjectManager
+import data.source.csv.user.CurrentUserProvider
 import domain.models.project.Project
 import domain.utils.PlanMateExceptions
 import domain.utils.ProjectExceptions
