+++ src/test/kotlin/data/repository/TaskRepositoryImplTest.kt
@@ -1,8 +1,8 @@
 package data.repository
 
 import data.mapper.toDomain
-import data.source.task.TaskExternalDataSource
-import data.source.task.TaskManager
+import data.source.csv.task.TaskExternalDataSource
+import data.source.csv.task.TaskManager
 import data.source.task.helperTaskDto
 import domain.repository.TaskRepository
 import domain.utils.TaskExceptions
