+++ src/test/kotlin/data/source/user/UserMongoDBDataSourceTest.kt
@@ -10,6 +10,8 @@ import com.mongodb.kotlin.client.coroutine.MongoCollection
 import com.mongodb.kotlin.client.coroutine.MongoDatabase
 import data.CopyCollectionIfDifferentToTest
 import data.dto.authentication.UserDto
+import data.source.csv.user.UserExternalDataSource
+import data.source.csv.user.UserMongoDBDataSource
 import data.source.mongoDb.MongoClientProvider
 import data.utils.PASSWORD
 import data.utils.USER_NAME
